Testing and debugging

+++ LacesAPI/LacesViewModel/Request/SearchRequest.cs
@@ -6,8 +6,8 @@ namespace LacesViewModel.Request
         public int SearchType { get; set; }
         public List<string> Keywords { get; set; }
         public List<string> Tags { get; set; }
-        public double MinPrice { get; set; }
-        public double MaxPrice { get; set; }
+        public decimal MinPrice { get; set; }
+        public decimal MaxPrice { get; set; }
         public string Size { get; set; }
         public string Brand { get; set; }
     }
